BitKeeper/etc/logging_ok:
  auto-union

+++ innobase/os/os0file.c
@@ -2053,8 +2053,8 @@ try_again:
 	ut_ad(buf);
 	ut_ad(n > 0);
 
-	low = offset;
-	high = offset_high;
+	low = (DWORD) offset;
+	high = (DWORD) offset_high;
 
 	/* Protect the seek / read operation with a mutex */
 	i = ((ulint) file) % OS_FILE_N_SEEK_MUTEXES;
@@ -2072,7 +2072,7 @@ try_again:
 	
         os_n_pending_reads++;
         
-	ret = ReadFile(file, buf, n, &len, NULL);
+	ret = ReadFile(file, buf, (DWORD) n, &len, NULL);
 
         os_n_pending_reads--;
 
@@ -2164,8 +2164,8 @@ try_again:
 	ut_ad(buf);
 	ut_ad(n > 0);
 
-	low = offset;
-	high = offset_high;
+	low = (DWORD) offset;
+	high = (DWORD) offset_high;
 
 	/* Protect the seek / read operation with a mutex */
 	i = ((ulint) file) % OS_FILE_N_SEEK_MUTEXES;
@@ -2183,7 +2183,7 @@ try_again:
 	
         os_n_pending_reads++;
         
-	ret = ReadFile(file, buf, n, &len, NULL);
+	ret = ReadFile(file, buf, (DWORD) n, &len, NULL);
 
         os_n_pending_reads--;
         
@@ -2258,8 +2258,8 @@ os_file_write(
 	ut_ad(buf);
 	ut_ad(n > 0);
 retry:
-	low = offset;
-	high = offset_high;
+	low = (DWORD) offset;
+	high = (DWORD) offset_high;
 	
 	/* Protect the seek / write operation with a mutex */
 	i = ((ulint) file) % OS_FILE_N_SEEK_MUTEXES;
@@ -2288,7 +2288,7 @@ retry:
 
         os_n_pending_writes++;
         
-	ret = WriteFile(file, buf, n, &len, NULL);
+	ret = WriteFile(file, buf, (DWORD) n, &len, NULL);
 
         os_n_pending_writes--;
 	
@@ -3279,7 +3279,7 @@ os_aio(
 #ifdef WIN_ASYNC_IO
 	ibool		retval;
 	BOOL		ret		= TRUE;
-	DWORD		len		= n;
+	DWORD		len		= (DWORD) n;
 	void*		dummy_mess1;
 	void*		dummy_mess2;
 	ulint		dummy_type;
@@ -4126,7 +4126,7 @@ loop:
 	if (os_n_file_reads == os_n_file_reads_old) {
 		avg_bytes_read = 0.0;
 	} else {
-		avg_bytes_read = os_bytes_read_since_printout /
+		avg_bytes_read = (double) os_bytes_read_since_printout /
 				(os_n_file_reads - os_n_file_reads_old);
 	}
